+++ flexmeasures/utils/config_defaults.py
@@ -72,8 +72,7 @@ class Config(object):
     SECURITY_TOKEN_MAX_AGE: int = 60 * 60 * 6  # six hours
     SECURITY_TRACKABLE: bool = False  # this is more in line with modern privacy law
     SECURITY_PASSWORD_SALT: str | None = None
-    FLEXMEASURES_SECRETS_ENCRYPTION_KEY: str | None = None
-    FLEXMEASURES_SECRETS_ENCRYPTION_KEY_ID: str = "default"
+    FLEXMEASURES_SECRETS_ENCRYPTION_KEYS: dict[str, str] | None = None
 
     # Two Factor Authentication
     SECURITY_TWO_FACTOR_ENABLED_METHODS = [
@@ -214,8 +213,8 @@ class Config(object):
 
 
 #  names of settings which cannot be None
-#  SECRET_KEY is also required but utils.app_utils.set_secret_key takes care of this better.
-#  SECURITY_TOTP_SECRETS is also required but utils.app_utils.set_totp_secrets takes care of this better.
+#  SECRET_KEY is also required but utils.secrets_utils.set_secret_key takes care of this better.
+#  SECURITY_TOTP_SECRETS is also required but utils.secrets_utils.set_totp_secrets takes care of this better.
 required: list[str] = ["SQLALCHEMY_DATABASE_URI"]
 
 #  settings whose absence should trigger a warning